add error message to catch too many open threads
When using many threads from the same process (the same limit of open
file handles), the plugin may crash. To identify the cause, the perf
tmam handler is extended by an error message if opening the cpu type
file fails (which is purely a sanity check anyways). The conditions when
this may lead to a crash and how to mitigate it are added to the README.